pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "JustAnotherExpenseManager"
version = "0.0.1"
description = "A Flask-based expense tracking and management application"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "paperplate"}
]
keywords = ["flask", "expense-tracker", "finance", "budgeting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Flask",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "flask>=3.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
]

[project.optional-dependencies]
# Testing dependencies
test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-flask>=1.3.0",
    "coverage>=7.3.2",
]

# Development dependencies
dev = [
    "flake8>=6.1.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]

# Performance testing
perf = [
    "pytest-benchmark>=4.0.0",
    "pytest-xdist>=3.5.0",
]

# Documentation
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
]

# All extras combined
all = [
    "JustAnotherExpenseManager[test,dev,perf,docs]",
]

[project.urls]
Homepage = "https://github.com/paperplate/JustAnotherExpenseManager"
Repository = "https://github.com/paperplate/JustAnotherExpenseManager"

[project.scripts]
JustAnotherExpenseManager = "JustAnotherExpenseManager:main"

[tool.setuptools]
packages = ["JustAnotherExpenseManager", "JustAnotherExpenseManager.routes", "JustAnotherExpenseManager.utils", "JustAnotherExpenseManager.models"]

[tool.setuptools.package-data]
"JustAnotherExpenseManager" = ["templates/*", "static/*", "static/css/*", "static/js/*"]

# Black configuration
[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | env
  | build
  | dist
  | migrations
  | __pycache__
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "venv/*", "env/*"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=JustAnotherExpenseManager",
    "--cov-report=html",
    "--cov-report=term-missing",
]

# Coverage configuration
[tool.coverage.run]
source = ["JustAnotherExpenseManager"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
